tests: fix fork--pidns-translation test on sparc and sparc64.
On sparc/sparc64, the value returned by syscall(__NR_fork) to the child
process is not 0 but the pid of the parent process in the init
namespace, so a simple check for non-zero is not a correct way
to distinguish the parent process from its child process.

* tests/fork--pidns-translation.c (fork_chain): When syscall(__NR_fork)
returns a non-zero value, do not assume blindly that it's the child process,
double check using waitpid.
(main): Call perror_msg_and_fail instead of perror.